requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-limiter>=3.5.0
sqlalchemy>=2.0.32
alembic>=1.13.0
psycopg2-binary>=2.9.10
celery>=5.4.0
redis>=5.0.1
google-generativeai>=0.4.0
twilio>=8.10.0
pydantic[email]>=2.9.2
pydantic-settings>=2.5.2
python-jose>=3.3.0
bcrypt>=4.1.2
python-dotenv>=1.0.0
bleach>=6.1.0
flask-cors>=4.0.0
requests>=2.31.0
pytest>=7.4.3
google-genai>=1.0.0
cryptography>=43.0.0
sentry-sdk[flask]>=2.16.0
